Atualização da função de atualizar atividade

+++ src/controllers/activityController.ts
@@ -4,6 +4,7 @@ import emailEvaluationResult from "../components/emailEvaluationResult";
 import checkActivityExistence from "../components/checkActivityExistence";
 import validateActivityInformation from "../components/validateActivityInformation";
 import activityCreateNotificationEmail from "../components/activityCreateNotificationEmail";
+import activityUpdateNotificationEmail from "../components/activityUpdateNotificationEmail";
 
 const prisma = new PrismaClient();
 
@@ -46,7 +47,7 @@ class ActivityController {
                 });
 
                 if (activity) {
-                    activityCreateNotificationEmail(activity)
+                    await activityCreateNotificationEmail(activity)
                     res.json({message: "Atividade cadastrada com sucesso", activity})
                     return;
                 }
@@ -101,28 +102,40 @@ class ActivityController {
         try {
             const { id } = req.params;
             const { name, activityGroup, activityType, workload, activityPeriod, placeOfCourse } = req.body;
-            const activity = await prisma.activity.findUnique({
-                where: { id: Number(id) },
-            });
-            if (!activity) {
-                res.status(404).json({ error: "Atividade não encontrado." });
+            const file = req.file;
+
+            if (!file || file.mimetype !== 'application/pdf') {
+                res.status(401).json({ message: 'Nenhum arquivo PDF enviado.' });
             } else {
-                if (activity.evaluation == "Deferida") {
-                    res.status(405).json({ error: "Não é possível atualizar a atividade deferida." });
+                const certificates = file.path;
+                const activity = await prisma.activity.findUnique({
+                    where: { id: Number(id) },
+                });
+                if (!activity) {
+                    res.status(404).json({ error: "Atividade não encontrado." });
                 } else {
-                    const activityUpdate = await prisma.activity.update({
-                        where: { id: Number(id) },
-                        data: {
-                            name, 
-                            activityGroup, 
-                            activityType, 
-                            workload,
-                            activityPeriod,
-                            placeOfCourse,
-                            certificate: "aa"
-                        },
-                    });
-                    res.json({ activityUpdate });
+                    if (activity.evaluation == "Deferida") {
+                        res.status(405).json({ error: "Não é possível atualizar a atividade deferida." });
+                    } else {
+                        const activityUpdate = await prisma.activity.update({
+                            where: { id: Number(id) },
+                            data: {
+                                name, 
+                                activityGroup, 
+                                activityType, 
+                                workload: Number(workload),
+                                activityPeriod,
+                                placeOfCourse,
+                                certificate: certificates,
+                                evaluation: "Em análise"
+                            },
+                        });
+                        if (activityUpdate) {
+                            await activityUpdateNotificationEmail(activityUpdate)
+                            res.json({message: "Atividade atualizada com sucesso", activityUpdate})
+                            return;
+                        }
+                    }
                 }
             }
         } catch (error) {
@@ -164,10 +177,14 @@ class ActivityController {
             if (!activity) {
                 res.status(404).json({ error: "Atividade não encontrado." });
             } else {
-                await prisma.activity.delete({
-                    where: { id: Number(id) },
-                });
-                res.json({ message: "Atividade excluída com sucesso." });
+                if(activity.evaluation == "Deferida"){
+                    res.status(402).json({ error: "Você não pode excluir uma atividade Deferida." });
+                } else {
+                    await prisma.activity.delete({
+                        where: { id: Number(id) },
+                    });
+                    res.json({ message: "Atividade excluída com sucesso." });
+                }
             }
         } catch (error) {
             res.status(500).json({ error: "Ocorreu um erro ao excluir a atividade." });
